@@ -1,7 +1,7 @@
-/**
+/*
  * MIT License
  *
- * Copyright (c) 2016 Azamshul Azizy
+ * Copyright (c) 2016-2025 Azamshul Azizy
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
  * associated documentation files (the "Software"), to deal in the Software without restriction,
@@ -20,24 +20,33 @@
  */
 package io.azam.ulidj;
 
+import org.junit.jupiter.api.Test;
+
 import java.security.SecureRandom;
+import java.time.Clock;
+import java.time.Duration;
+import java.time.Instant;
+import java.time.ZoneId;
+import java.time.ZoneOffset;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Random;
+import java.util.concurrent.ThreadLocalRandom;
 
-import org.junit.Test;
-import org.junit.function.ThrowingRunnable;
-
+import static io.azam.ulidj.TestUtils.assertEntropyIsIncremented;
 import static io.azam.ulidj.TestUtils.incrementBytes;
 import static io.azam.ulidj.ULIDTest.FILLED_ENTROPY;
+import static io.azam.ulidj.ULIDTest.TEST_CLOCK;
 import static io.azam.ulidj.ULIDTest.TEST_RANDOM;
-import static org.junit.Assert.assertArrayEquals;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertThrows;
-import static org.junit.Assert.assertTrue;
+import static io.azam.ulidj.ULIDTest.TEST_TIMESTAMP;
+import static io.azam.ulidj.ULIDTest.ZERO_ENTROPY;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * Test class for {@link io.azam.ulidj.MonotonicULID}
@@ -51,8 +60,33 @@ public class MonotonicULIDTest {
     assertNotNull(new MonotonicULID());
     assertNotNull(new MonotonicULID(new Random()));
     assertNotNull(new MonotonicULID(new SecureRandom()));
-    Random nullRandom = null;
-    assertNotNull(new MonotonicULID(nullRandom));
+    assertNotNull(new MonotonicULID(ThreadLocalRandom.current()));
+    assertThrows(IllegalArgumentException.class, () -> {
+      Random random = null;
+      new MonotonicULID(random);
+    });
+    assertNotNull(new MonotonicULID(Clock.systemDefaultZone()));
+    assertNotNull(new MonotonicULID(Clock.fixed(Instant.ofEpochMilli(0L), ZoneId.systemDefault())));
+    assertThrows(IllegalArgumentException.class, () -> {
+      Clock clock = null;
+      new MonotonicULID(clock);
+    });
+    assertNotNull(new MonotonicULID(Clock.systemDefaultZone(), new Random()));
+    assertThrows(IllegalArgumentException.class, () -> {
+      Random random = null;
+      Clock clock = null;
+      new MonotonicULID(clock, random);
+    });
+    assertThrows(IllegalArgumentException.class, () -> {
+      Random random = new Random();
+      Clock clock = null;
+      new MonotonicULID(clock, random);
+    });
+    assertThrows(IllegalArgumentException.class, () -> {
+      Random random = null;
+      Clock clock = Clock.systemDefaultZone();
+      new MonotonicULID(clock, random);
+    });
   }
 
   @Test
@@ -77,30 +111,393 @@ public class MonotonicULIDTest {
     ULID value = ulid.generateULID();
     assertNotNull(value);
     assertEquals(ULID.class, value.getClass());
-    assertNotNull("ULID instance string value should not be null", value.toString());
-    assertTrue("ULID instance string value must be valid", ULID.isValid(value.toString()));
-    assertNotNull("ULID instance binary value should not be null", value.toBinary());
-    assertTrue("ULID instance binary value must be valid", ULID.isValidBinary(value.toBinary()));
+    assertNotNull(value.toString(), "ULID instance string value should not be null");
+    assertTrue(ULID.isValid(value.toString()), "ULID instance string value must be valid");
+    assertNotNull(value.toBinary(), "ULID instance binary value should not be null");
+    assertTrue(ULID.isValidBinary(value.toBinary()), "ULID instance binary value must be valid");
   }
 
   @Test
   public void testGenerateExternalRandom() {
     MonotonicULID ulid = new MonotonicULID(TEST_RANDOM);
     String value = ulid.generate();
-    assertNotNull("ULID value should not be null", value);
-    assertTrue("ULID value must be valid", ULID.isValid(value));
-    assertArrayEquals("ULID entropy should be filled with the provided random", FILLED_ENTROPY,
-        ULID.getEntropy(value));
+    assertNotNull(value, "ULID value should not be null");
+    assertTrue(ULID.isValid(value), "ULID value must be valid");
+    assertArrayEquals(FILLED_ENTROPY, ULID.getEntropy(value),
+        "ULID entropy should be filled with the provided random");
   }
 
   @Test
   public void testGenerateBinaryExternalRandom() {
     MonotonicULID ulid = new MonotonicULID(TEST_RANDOM);
     byte[] value = ulid.generateBinary();
-    assertNotNull("Binary ULID value should not be null", value);
-    assertTrue("Binary ULID value must be valid", ULID.isValidBinary(value));
-    assertArrayEquals("Binary ULID entropy should be filled with the provided random",
-        FILLED_ENTROPY, ULID.getEntropyBinary(value));
+    assertNotNull(value, "Binary ULID value should not be null");
+    assertTrue(ULID.isValidBinary(value), "Binary ULID value must be valid");
+    assertArrayEquals(FILLED_ENTROPY, ULID.getEntropyBinary(value),
+        "Binary ULID entropy should be filled with the provided random");
+  }
+
+  @Test
+  public void testGenerateULIDExternalRandom() {
+    MonotonicULID ulid = new MonotonicULID(TEST_RANDOM);
+    ULID value = ulid.generateULID();
+    assertNotNull(value, "ULID value should not be null");
+    assertArrayEquals(FILLED_ENTROPY, value.getEntropy(),
+        "Binary ULID entropy should be filled with the provided random");
+  }
+
+  @Test
+  public void testGenerateExternalClock() {
+    MonotonicULID ulid = new MonotonicULID(TEST_CLOCK);
+    String value = ulid.generate();
+    assertNotNull(value, "ULID value should not be null");
+    assertTrue(ULID.isValid(value), "ULID value must be valid");
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestamp(value),
+        "ULID timestamp should be filled with the provided clock");
+  }
+
+  @Test
+  public void testGenerateBinaryExternalClock() {
+    MonotonicULID ulid = new MonotonicULID(TEST_CLOCK);
+    byte[] value = ulid.generateBinary();
+    assertNotNull(value, "Binary ULID value should not be null");
+    assertTrue(ULID.isValidBinary(value), "Binary ULID value must be valid");
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestampBinary(value),
+        "Binary ULID timestamp should be filled with the provided clock");
+  }
+
+  @Test
+  public void testGenerateULIDExternalClock() {
+    MonotonicULID ulid = new MonotonicULID(TEST_CLOCK);
+    ULID value = ulid.generateULID();
+    assertNotNull(value, "ULID value should not be null");
+    assertEquals(TEST_TIMESTAMP, value.getTimestamp(),
+        "Binary ULID timestamp should be filled with the provided clock");
+  }
+
+  @Test
+  public void testGenerateExternalRandomAndClock() {
+    MonotonicULID ulid = new MonotonicULID(TEST_CLOCK, TEST_RANDOM);
+    String value = ulid.generate();
+    assertNotNull(value, "ULID value should not be null");
+    assertTrue(ULID.isValid(value), "ULID value must be valid");
+    assertArrayEquals(FILLED_ENTROPY, ULID.getEntropy(value),
+        "ULID entropy should be filled with the provided random");
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestamp(value),
+        "ULID timestamp should be filled with the provided clock");
+  }
+
+  @Test
+  public void testGenerateBinaryExternalClockAndRandom() {
+    MonotonicULID ulid = new MonotonicULID(TEST_CLOCK, TEST_RANDOM);
+    byte[] value = ulid.generateBinary();
+    assertNotNull(value, "Binary ULID value should not be null");
+    assertTrue(ULID.isValidBinary(value), "Binary ULID value must be valid");
+    assertArrayEquals(FILLED_ENTROPY, ULID.getEntropyBinary(value),
+        "Binary ULID entropy should be filled with the provided random");
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestampBinary(value),
+        "Binary ULID timestamp should be filled with the provided clock");
+  }
+
+  @Test
+  public void testGenerateULIDExternalClockAndRandom() {
+    MonotonicULID ulid = new MonotonicULID(TEST_CLOCK, TEST_RANDOM);
+    ULID value = ulid.generateULID();
+    assertNotNull(value, "Binary ULID value should not be null");
+    assertArrayEquals(FILLED_ENTROPY, value.getEntropy(),
+        "Binary ULID entropy should be filled with the provided random");
+    assertEquals(TEST_TIMESTAMP, value.getTimestamp(),
+        "Binary ULID timestamp should be filled with the provided clock");
+  }
+
+  @Test
+  public void testGenerateConcurrent() {
+    Instant instant = Instant.ofEpochMilli(TEST_TIMESTAMP);
+    MutableFixedClock clock = new MutableFixedClock(instant, ZoneOffset.UTC);
+    MonotonicULID ulid = new MonotonicULID(clock, new FixedRandom(ZERO_ENTROPY));
+    // Base
+    String ulid1 = ulid.generate();
+    assertTrue(ULID.isValid(ulid1));
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestamp(ulid1));
+    assertArrayEquals(ZERO_ENTROPY, ULID.getEntropy(ulid1));
+    // Same timestamp as base, entropy should be incremented
+    String ulid2 = ulid.generate();
+    assertTrue(ULID.isValid(ulid2));
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestamp(ulid2));
+    assertEntropyIsIncremented(ulid1, ulid2);
+    // Fast forward clock 1ms to the future
+    clock.tick(Duration.ofMillis(1));
+    assertEquals(TEST_TIMESTAMP + 1, clock.instant().toEpochMilli());
+    // Base timestamp + 1ms, entropy should reset
+    String ulid3 = ulid.generate();
+    assertTrue(ULID.isValid(ulid3));
+    assertEquals(TEST_TIMESTAMP + 1, ULID.getTimestamp(ulid3));
+    assertArrayEquals(ZERO_ENTROPY, ULID.getEntropy(ulid3));
+  }
+
+  @Test
+  public void testGenerateBinaryConcurrent() {
+    Instant instant = Instant.ofEpochMilli(TEST_TIMESTAMP);
+    MutableFixedClock clock = new MutableFixedClock(instant, ZoneOffset.UTC);
+    MonotonicULID ulid = new MonotonicULID(clock, new FixedRandom(ZERO_ENTROPY));
+    // Base
+    byte[] ulid1 = ulid.generateBinary();
+    assertTrue(ULID.isValidBinary(ulid1));
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestampBinary(ulid1));
+    assertArrayEquals(ZERO_ENTROPY, ULID.getEntropyBinary(ulid1));
+    // Same timestamp as base, entropy should be incremented
+    byte[] ulid2 = ulid.generateBinary();
+    assertTrue(ULID.isValidBinary(ulid2));
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestampBinary(ulid2));
+    assertEntropyIsIncremented(ulid1, ulid2);
+    // Fast forward clock 1ms to the future
+    clock.tick(Duration.ofMillis(1));
+    assertEquals(TEST_TIMESTAMP + 1, clock.instant().toEpochMilli());
+    // Base timestamp + 1ms, entropy should reset
+    byte[] ulid3 = ulid.generateBinary();
+    assertTrue(ULID.isValidBinary(ulid3));
+    assertEquals(TEST_TIMESTAMP + 1, ULID.getTimestampBinary(ulid3));
+    assertArrayEquals(ZERO_ENTROPY, ULID.getEntropyBinary(ulid3));
+  }
+
+  @Test
+  public void testGenerateConcurrentEntropyOverflow() {
+    Instant instant = Instant.ofEpochMilli(TEST_TIMESTAMP);
+    MutableFixedClock clock = new MutableFixedClock(instant, ZoneOffset.UTC);
+    MonotonicULID ulid = new MonotonicULID(clock, new FixedRandom(FILLED_ENTROPY));
+    // Base
+    String ulid1 = ulid.generate();
+    assertTrue(ULID.isValid(ulid1));
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestamp(ulid1));
+    assertArrayEquals(FILLED_ENTROPY, ULID.getEntropy(ulid1));
+    // Entropy overflow, should throw exception
+    assertThrows(IllegalStateException.class, () -> {
+      String ulid2 = ulid.generate();
+    }, "Entropy is filled, so the next generation in the same timestamp should fail");
+    // Entropy overflow, should throw exception
+    assertThrows(IllegalStateException.class, () -> {
+      String ulid2 = ulid.generate();
+    }, "Subsequent generations should also fail");
+  }
+
+  @Test
+  public void testGenerateBinaryConcurrentEntropyOverflow() {
+    Instant instant = Instant.ofEpochMilli(TEST_TIMESTAMP);
+    MutableFixedClock clock = new MutableFixedClock(instant, ZoneOffset.UTC);
+    MonotonicULID ulid = new MonotonicULID(clock, new FixedRandom(FILLED_ENTROPY));
+    // Base
+    byte[] ulid1 = ulid.generateBinary();
+    assertTrue(ULID.isValidBinary(ulid1));
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestampBinary(ulid1));
+    assertArrayEquals(FILLED_ENTROPY, ULID.getEntropyBinary(ulid1));
+    // Entropy overflow, should throw exception
+    assertThrows(IllegalStateException.class, () -> {
+      byte[] ulid2 = ulid.generateBinary();
+    }, "Entropy is filled, so the next generation in the same timestamp should fail");
+    // Entropy overflow, should throw exception
+    assertThrows(IllegalStateException.class, () -> {
+      byte[] ulid2 = ulid.generateBinary();
+    }, "Subsequent generations should also fail");
+  }
+
+  @Test
+  public void testGenerateTimestampUnderflow() {
+    Instant instant = Instant.ofEpochMilli(ULID.MIN_TIME);
+    MutableFixedClock clock = new MutableFixedClock(instant, ZoneOffset.UTC);
+    MonotonicULID ulid = new MonotonicULID(clock, new FixedRandom(ZERO_ENTROPY));
+    // Base
+    String ulid1 = ulid.generate();
+    assertTrue(ULID.isValid(ulid1));
+    assertEquals(ULID.MIN_TIME, ULID.getTimestamp(ulid1));
+    // Default last timestamp is ULID.MIN_TIME with last entropy of ULID.ZERO_ENTROPY
+    // so this is should be incremented
+    assertArrayEquals(incrementBytes(ZERO_ENTROPY), ULID.getEntropy(ulid1));
+    // Reveerse clock 1ms to the past
+    clock.tick(Duration.ofMillis(-1));
+    assertEquals(ULID.MIN_TIME - 1, clock.instant().toEpochMilli());
+    // Timestamp is out of range (lesser than ULID.MIN_TIME), should throw exception
+    assertThrows(IllegalStateException.class, () -> {
+      String ulid2 = ulid.generate();
+    }, "Timestamp is lesser than ULID.MIN_TIME, so the next generation should fail");
+    assertThrows(IllegalStateException.class, () -> {
+      String ulid2 = ulid.generate();
+    }, "Subsequent generations should also fail");
+  }
+
+  @Test
+  public void testGenerateBinaryTimestampUnderflow() {
+    Instant instant = Instant.ofEpochMilli(ULID.MIN_TIME);
+    MutableFixedClock clock = new MutableFixedClock(instant, ZoneOffset.UTC);
+    MonotonicULID ulid = new MonotonicULID(clock, new FixedRandom(ZERO_ENTROPY));
+    // Base
+    byte[] ulid1 = ulid.generateBinary();
+    assertTrue(ULID.isValidBinary(ulid1));
+    assertEquals(ULID.MIN_TIME, ULID.getTimestampBinary(ulid1));
+    // Default last timestamp is ULID.MIN_TIME with last entropy of ULID.ZERO_ENTROPY
+    // so this is should be incremented
+    assertArrayEquals(incrementBytes(ZERO_ENTROPY), ULID.getEntropyBinary(ulid1));
+    // Reveerse clock 1ms to the past
+    clock.tick(Duration.ofMillis(-1));
+    assertEquals(ULID.MIN_TIME - 1, clock.instant().toEpochMilli());
+    // Timestamp is out of range (lesser than ULID.MIN_TIME), should throw exception
+    assertThrows(IllegalStateException.class, () -> {
+      byte[] ulid2 = ulid.generateBinary();
+    }, "Timestamp is lesser than ULID.MIN_TIME, so the next generation should fail");
+    assertThrows(IllegalStateException.class, () -> {
+      byte[] ulid2 = ulid.generateBinary();
+    }, "Subsequent generations should also fail");
+  }
+
+  @Test
+  public void testGenerateTimestampOverflow() {
+    Instant instant = Instant.ofEpochMilli(ULID.MAX_TIME);
+    MutableFixedClock clock = new MutableFixedClock(instant, ZoneOffset.UTC);
+    MonotonicULID ulid = new MonotonicULID(clock, new FixedRandom(ZERO_ENTROPY));
+    // Base
+    String ulid1 = ulid.generate();
+    assertTrue(ULID.isValid(ulid1));
+    assertEquals(ULID.MAX_TIME, ULID.getTimestamp(ulid1));
+    assertArrayEquals(ZERO_ENTROPY, ULID.getEntropy(ulid1));
+    // Fast forward clock 1ms to the future
+    clock.tick(Duration.ofMillis(1));
+    assertEquals(ULID.MAX_TIME + 1, clock.instant().toEpochMilli());
+    // Timestamp is out of range (greater than ULID.MAX_TIME), should throw exception
+    assertThrows(IllegalStateException.class, () -> {
+      String ulid2 = ulid.generate();
+    }, "Timestamp is greater than ULID.MAX_TIME, so the next generation should fail");
+    assertThrows(IllegalStateException.class, () -> {
+      String ulid2 = ulid.generate();
+    }, "Subsequent generations should also fail");
+  }
+
+  @Test
+  public void testGenerateBinaryTimestampOverflow() {
+    Instant instant = Instant.ofEpochMilli(ULID.MAX_TIME);
+    MutableFixedClock clock = new MutableFixedClock(instant, ZoneOffset.UTC);
+    MonotonicULID ulid = new MonotonicULID(clock, new FixedRandom(ZERO_ENTROPY));
+    // Base
+    byte[] ulid1 = ulid.generateBinary();
+    assertTrue(ULID.isValidBinary(ulid1));
+    assertEquals(ULID.MAX_TIME, ULID.getTimestampBinary(ulid1));
+    assertArrayEquals(ZERO_ENTROPY, ULID.getEntropyBinary(ulid1));
+    // Fast forward clock 1ms to the future
+    clock.tick(Duration.ofMillis(1));
+    assertEquals(ULID.MAX_TIME + 1, clock.instant().toEpochMilli());
+    // Timestamp is out of range (greater than ULID.MAX_TIME), should throw exception
+    assertThrows(IllegalStateException.class, () -> {
+      byte[] ulid2 = ulid.generateBinary();
+    }, "Timestamp is greater than ULID.MAX_TIME, so the next generation should fail");
+    assertThrows(IllegalStateException.class, () -> {
+      byte[] ulid2 = ulid.generateBinary();
+    }, "Subsequent generations should also fail");
+  }
+
+  @Test
+  public void testGenerateClockMovedBackwards() {
+    Instant instant = Instant.ofEpochMilli(TEST_TIMESTAMP);
+    MutableFixedClock clock = new MutableFixedClock(instant, ZoneOffset.UTC);
+    MonotonicULID ulid = new MonotonicULID(clock, new FixedRandom(ZERO_ENTROPY));
+    // Base
+    String ulid1 = ulid.generate();
+    assertTrue(ULID.isValid(ulid1));
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestamp(ulid1));
+    assertArrayEquals(ZERO_ENTROPY, ULID.getEntropy(ulid1));
+    // Reverse clock 1ms to the past
+    clock.tick(Duration.ofMillis(-1));
+    assertEquals(TEST_TIMESTAMP - 1, clock.instant().toEpochMilli());
+    // Timestamp is lesser than MonotonicULID.lastTimestamp,
+    // should use MonotonicULID.lastTimestamp and increment entropy
+    String ulid2 = ulid.generate();
+    assertTrue(ULID.isValid(ulid2));
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestamp(ulid2));
+    assertEntropyIsIncremented(ulid1, ulid2);
+    // Timestamp is still lesser than MonotonicULID.lastTimestamp,
+    // should use MonotonicULID.lastTimestamp and increment entropy
+    String ulid3 = ulid.generate();
+    assertTrue(ULID.isValid(ulid3));
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestamp(ulid3));
+    assertEntropyIsIncremented(ulid2, ulid3);
+  }
+
+  @Test
+  public void testGenerateBinaryClockMovedBackwards() {
+    Instant instant = Instant.ofEpochMilli(TEST_TIMESTAMP);
+    MutableFixedClock clock = new MutableFixedClock(instant, ZoneOffset.UTC);
+    MonotonicULID ulid = new MonotonicULID(clock, new FixedRandom(ZERO_ENTROPY));
+    // Base
+    byte[] ulid1 = ulid.generateBinary();
+    assertTrue(ULID.isValidBinary(ulid1));
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestampBinary(ulid1));
+    assertArrayEquals(ZERO_ENTROPY, ULID.getEntropyBinary(ulid1));
+    // Reverse clock 1ms to the past
+    clock.tick(Duration.ofMillis(-1));
+    assertEquals(TEST_TIMESTAMP - 1, clock.instant().toEpochMilli());
+    // Timestamp is lesser than MonotonicULID.lastTimestamp,
+    // should use MonotonicULID.lastTimestamp and increment entropy
+    byte[] ulid2 = ulid.generateBinary();
+    assertTrue(ULID.isValidBinary(ulid2));
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestampBinary(ulid2));
+    assertEntropyIsIncremented(ulid1, ulid2);
+    // Timestamp is still lesser than MonotonicULID.lastTimestamp,
+    // should use MonotonicULID.lastTimestamp and increment entropy
+    byte[] ulid3 = ulid.generateBinary();
+    assertTrue(ULID.isValidBinary(ulid3));
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestampBinary(ulid3));
+    assertEntropyIsIncremented(ulid2, ulid3);
+  }
+
+  @Test
+  public void testGenerateClockMovedForwards() {
+    Instant instant = Instant.ofEpochMilli(TEST_TIMESTAMP);
+    MutableFixedClock clock = new MutableFixedClock(instant, ZoneOffset.UTC);
+    MonotonicULID ulid = new MonotonicULID(clock, new FixedRandom(ZERO_ENTROPY));
+    // Base
+    String ulid1 = ulid.generate();
+    assertTrue(ULID.isValid(ulid1));
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestamp(ulid1));
+    assertArrayEquals(ZERO_ENTROPY, ULID.getEntropy(ulid1));
+    // Fast forward clock 1ms to the future
+    clock.tick(Duration.ofMillis(1));
+    assertEquals(TEST_TIMESTAMP + 1, clock.instant().toEpochMilli());
+    // Timestamp is greater than MonotonicULID.lastTimestamp,
+    // should use generate new entropy
+    String ulid2 = ulid.generate();
+    assertTrue(ULID.isValid(ulid2));
+    assertEquals(TEST_TIMESTAMP + 1, ULID.getTimestamp(ulid2));
+    assertArrayEquals(ZERO_ENTROPY, ULID.getEntropy(ulid2));
+    // Timestamp is same MonotonicULID.lastTimestamp,
+    // should increment entropy
+    String ulid3 = ulid.generate();
+    assertTrue(ULID.isValid(ulid3));
+    assertEquals(TEST_TIMESTAMP + 1, ULID.getTimestamp(ulid3));
+    assertEntropyIsIncremented(ulid2, ulid3);
+  }
+
+  @Test
+  public void testGenerateBinaryClockMovedForwards() {
+    Instant instant = Instant.ofEpochMilli(TEST_TIMESTAMP);
+    MutableFixedClock clock = new MutableFixedClock(instant, ZoneOffset.UTC);
+    MonotonicULID ulid = new MonotonicULID(clock, new FixedRandom(ZERO_ENTROPY));
+    // Base
+    byte[] ulid1 = ulid.generateBinary();
+    assertTrue(ULID.isValidBinary(ulid1));
+    assertEquals(TEST_TIMESTAMP, ULID.getTimestampBinary(ulid1));
+    assertArrayEquals(ZERO_ENTROPY, ULID.getEntropyBinary(ulid1));
+    // Fast forward clock 1ms to the future
+    clock.tick(Duration.ofMillis(1));
+    assertEquals(TEST_TIMESTAMP + 1, clock.instant().toEpochMilli());
+    // Timestamp is greater than MonotonicULID.lastTimestamp,
+    // should use generate new entropy
+    byte[] ulid2 = ulid.generateBinary();
+    assertTrue(ULID.isValidBinary(ulid2));
+    assertEquals(TEST_TIMESTAMP + 1, ULID.getTimestampBinary(ulid2));
+    assertArrayEquals(ZERO_ENTROPY, ULID.getEntropyBinary(ulid2));
+    // Timestamp is same MonotonicULID.lastTimestamp,
+    // should increment entropy
+    byte[] ulid3 = ulid.generateBinary();
+    assertTrue(ULID.isValidBinary(ulid3));
+    assertEquals(TEST_TIMESTAMP + 1, ULID.getTimestampBinary(ulid3));
+    assertEntropyIsIncremented(ulid2, ulid3);
   }
 
   @Test
@@ -203,13 +600,10 @@ public class MonotonicULIDTest {
         (byte) 0xff, (byte) 0xff, (byte) 0xff, (byte) 0xff, (byte) 0xff, //
         (byte) 0xff, (byte) 0xff, (byte) 0xff, (byte) 0xff, (byte) 0xff //
     }));
-    assertThrows(IllegalStateException.class, new ThrowingRunnable() {
-      @Override
-      public void run() {
-        List<String> values = new ArrayList<String>();
-        for (int i = 0; i < 1000000; i++) {
-          values.add(ulid.generate());
-        }
+    assertThrows(IllegalStateException.class, () -> {
+      List<String> values = new ArrayList<String>();
+      for (int i = 0; i < 1000000; i++) {
+        values.add(ulid.generate());
       }
     });
   }
@@ -222,13 +616,10 @@ public class MonotonicULIDTest {
         (byte) 0xff, (byte) 0xff, (byte) 0xff, (byte) 0xff, (byte) 0xff, //
         (byte) 0xff, (byte) 0xff, (byte) 0xff, (byte) 0xff, (byte) 0xff //
     }));
-    assertThrows(IllegalStateException.class, new ThrowingRunnable() {
-      @Override
-      public void run() {
-        List<byte[]> values = new ArrayList<byte[]>();
-        for (int i = 0; i < 1000000; i++) {
-          values.add(ulid.generateBinary());
-        }
+    assertThrows(IllegalStateException.class, () -> {
+      List<byte[]> values = new ArrayList<byte[]>();
+      for (int i = 0; i < 1000000; i++) {
+        values.add(ulid.generateBinary());
       }
     });
   }
